+++ src/routes/+page.server.ts
@@ -34,7 +34,9 @@ export const load: PageServerLoad = async () => {
 			preloadAhead: config.preloadAhead,
 			preloadBehind: config.preloadBehind,
 			autoAdvance: config.autoAdvance,
-			posters: config.dataDir !== ''
+			posters: config.dataDir !== '',
+			debugPlayback: config.debugPlayback,
+			buildSha: config.buildSha
 		}
 	};
 };